+++ fragments/default_theme/mform_default.php
@@ -2,16 +2,17 @@
 
 switch ($this->type) {
     case 'default':
-        $this->labelColClass = 'col-sm-2';
+        $this->labelColClass = 'col-sm-2 control-label';
         $this->formItemColClass = 'col-sm-10';
         break;
     case 'default_full':
+    case 'default_custom_full':
         $this->labelColClass = 'col-sm-12';
         $this->formItemColClass = 'col-sm-12';
         break;
 }
 
 echo '<div class="form-group">
-        <div class="' . $this->labelColClass . ' control-label">' . $this->label . $this->infoTooltip . $this->infoCollapseButton . '</div>
+        <div class="' . $this->labelColClass . '">' . $this->label . $this->infoTooltip . $this->infoCollapseButton . '</div>
         <div class="' . $this->formItemColClass . '">' . $this->element . '</div>
       </div>' . $this->infoCollapse;